feat: prevent duplicate keywords
Introduce a `unique` prop to `TagInput` to prevent duplicate keywords from being added.

This change adds a `unique` prop to the `TagInput` component, which when set to `true`, will prevent duplicate keywords from being added to the list of keywords. This can be useful in scenarios where it's important to ensure that all keywords in the list are unique.

The following changes were made to implement this feature:

- Added a `unique` prop to the `TagInput` component.
- Updated the `onKeywordsChange` handler to check if the `unique` prop is set to `true`.
- If the `unique` prop is set to `true`, the new keyword is added to a set of existing keywords to ensure uniqueness.
- The updated set of keywords is then used to update the state of the `TagInput` component.